pyproject.toml
[build-system]
requires = ["setuptools>=69.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentbox"
version = "0.1.0"
description = "AgentBox platform for deploying AI agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.2.1",
  "sqlalchemy[asyncio]>=2.0.0",
  "asyncpg>=0.29.0",
  "alembic>=1.13.0",
  "pyyaml>=6.0.1",
  "litellm>=1.0.0",
  "httpx>=0.27.0",
  "google-api-python-client>=2.0.0",
  "google-auth>=2.0.0",
  "redis>=5.0.0",
  "celery>=5.4.0",
  "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "httpx>=0.27.0",
  "ruff>=0.4.0",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[project.scripts]
agentbox = "src.cli:cli"
